Release v0.5.12: Looped Transformer GPU graph fusion, clippy fixes
- GPU acceleration for Nanbeige 4.2 (Metal): fuse both loop iterations
  into a single graph submission with on-device loop_final_norm
  (encode_loop_norm: RMS norm + blit). 5.6 → 18.6 tok/s decode (3.3×).
- GPU graph prefill for looped models: graph_prefill_preferred() = true
  for loop_final_norm archs on macOS. TTFT 5.2 → 1.8 s (2.9×).
- Fix GPU loop_final_norm insertion: continue after q1_graph_gpu /
  chunk_run_gpu no longer skips the per-loop norm.
- Clippy: div_ceil, needless_borrow, needless_range_loop, missing
  struct fields in tests (num_loops, loop_final_norm).

+++ crates/cortiq-engine/src/loader.rs
@@ -434,7 +434,7 @@ impl Pipeline {
                         value_head_dim: lc.value_head_dim,
                         conv_kernel: need(arch.linear_conv_kernel_dim, "linear_conv_kernel_dim")?,
                         hidden_size: arch.hidden_size,
-                        rms_eps: arch.rms_norm_eps as f64,
+                        rms_eps: arch.rms_norm_eps,
                     });
                 }
                 other => {
